Moved ShoppingCartContext to a src/contexts/ folder to remove circular dependency warning
